Move biome migrations into top level directory
All migrations will be moved to this file to avoid
possible conflicts when running migrations multiple
times.
