+++ libsplinter/src/biome/mod.rs
@@ -31,8 +31,6 @@ pub mod credentials;
 #[cfg(feature = "biome-key-management")]
 pub mod key_management;
 
-pub mod migrations;
-
 #[cfg(feature = "biome-notifications")]
 pub mod notifications;
